Move folders and add icons
pictures dont work yet nor buttons
